Finishing overlaying object Tune, creating data structure array list, implementing an interface
Finsished up all get, sets, and other parts of the overlaying object Tune. Included initializing the tune or by calling all the underlaying objects for their user input.

Created the beginning of the overlaying data structure to be used for the arrayList containing the Tunes. Based off of a interface and with the object type being Tune.